# lib/api-zod/src/generated/types/dailyEarning.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LINESETU Queue Management API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyEarning {
  date: string;
  totalTokens: number;
  tokensNormal?: number;
  tokensEmergency?: number;
  earned: number;
}