# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LINESETU Queue Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Patient {
  id: string;
  name: string;
  phone: string;
  profilePhoto?: string;
  isNew?: boolean;
}